syncbase: Client should not be an Application.
I realized last night that I was doing the wrong thing by making the
Clients extend Application.  They should not do that.

The scripts in bin/ and test/ now create their own Application, and pass
it to the Client.

Also, non-PHONY tasks that used to depend on "gen-mojom" now have a
prereq of the .mojom.go or .mojom.dart files that they actually depend
on.  Since "gen-mojom" is PHONY, it was causing extra work in non-PHONY
tasks that depended on it.

Change-Id: Ib3f693ffc0da427b04f3da9e726cbd86ca01fbae