syncbase: Echo and syncbase tests run, pass, and exit reliably.
The biggest issue was that when --enable-multiprocess is set (as it must
be for Go/Mojo apps), closing the main application causes mojo_shell to
crash.  Not closing an application, however, causes mojo_shell to hang.

The solution is to wait for the test reporter to finish, then close the
application.  mojo_shell still crashes, but the test harness is able to
find the test reporter output and exit with the correct status code.

Calling a "Quit" method in Syncbase made no difference.  I got rid of
the "Quit" method in the echo_server, since we weren't using it anyways.

This brought a tear to my eye:

    Running dart/test/echo_test.dart.... Succeeded
    Running dart/test/syncbase_test.dart.... Succeeded

Change-Id: Idd25d5369cd5dc19c572d2e7150b0e26253ab75d