mojo/syncbase: Fixing handle leaks and forcing strict mode for tests.
We were leaking mojo handles for proxies that
are created for Watch, Exec and Scan streams.
These proxies are created so Go code can call
into Dart and the handles are not tied to the
client proxy that is used to call Go code from
Dart and therefore they need to be tracked and
closed manually and neither mojo nor generated
mojom code knows about their existence.

This CL introduces a UnclosedStubsManager which
methods like Watch, Scan, Exec add their stubs
to and remove it when they finish. Any stubs that
are still open when SyncbaseClient.close() is called
are then closed preventing handle leaks.

This CL also refactors and bundles Proxy and
UnclosedStubsManager into a single ClientContext
that is passed from client to layers that need it.

NOTE: Currently we can not actually test for leaks
due to the bug https://github.com/domokit/mojo/issues/487
which is crashing the test runner when `app.close` is
called. Without `app.close()` being called before checking
for leaks, some handles will always leak. For now I have
verified leaks are down to 2 which are unrelated to our code
and are due to `app.close()` not being called but this
is not good, and hopefully that bug gets fixes soon so
we can enable the leak check.

Also note that `reportLeakedHandles` does not work in
non-strict mode so we were never actually testing for
leaks.

Change-Id: Ie00442bec5156fd1e8ecb8afb90a3479d66c7189