mojo/syncbase: Add Local Publish rule
This adds a new rule called local-publish to the Makefile.

Users of syncbase that want to add debug logs, try out a fix, or
use the latest (but unpublished) version of syncbase can do so with
this rule.

Note: While this currently lets you pick any version of Mojo to
build with, that will change once
https://vanadium-review.googlesource.com/#/c/17692/
is submitted.

Change-Id: I9a99ed6a5cafd90563a682181610a64ab809da0c