mojo/syncbase: Retry Syncgroup Join
This bug appeared soon after submitting my benchmark.

When running on Android, it looks like the syncgroup join occasionally
fails. It is not clear to me why this is the case, but in order to
avoid random crashes, I have switched to using a while loop + try
catch block.

This seems to stabilize the benchmark, but... it is still a bug.
https://github.com/vanadium/issues/issues/1162

Croupier does not run into this, so it may be a recent change that
is causing this. Or it may be related to the syncgroup create bug.

Change-Id: I66e454c9164c50a90d11c848e226277b7c015813